requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pytz>=2023.3
pytest>=7.0.0
freezegun>=1.5.0
spacy>=3.8.0
numpy>=1.24.0
